+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -26234,7 +26234,7 @@ $GD_Completers = {
         # Amazon.GuardDuty.UsageStatisticType
         "Get-GDUsageStatistic/UsageStatisticType"
         {
-            $v = "SUM_BY_ACCOUNT","SUM_BY_DATA_SOURCE","SUM_BY_RESOURCE","TOP_RESOURCES"
+            $v = "SUM_BY_ACCOUNT","SUM_BY_DATA_SOURCE","SUM_BY_FEATURES","SUM_BY_RESOURCE","TOP_RESOURCES"
             break
         }