+++ modules/AWSPowerShell/Cmdlets/SageMakerRuntime/Basic/Invoke-SMREndpointAsync-Cmdlet.cs
@@ -151,7 +151,7 @@ namespace Amazon.PowerShell.Cmdlets.SMR
         /// <summary>
         /// <para>
         /// <para>Maximum amount of time in seconds a request can be processed before it is marked as
-        /// expired.</para>
+        /// expired. The default is 15 minutes, or 900 seconds.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -162,7 +162,8 @@ namespace Amazon.PowerShell.Cmdlets.SMR
         #region Parameter RequestTTLSecond
         /// <summary>
         /// <para>
-        /// <para>Maximum age in seconds a request can be in the queue before it is marked as expired.</para>
+        /// <para>Maximum age in seconds a request can be in the queue before it is marked as expired.
+        /// The default is 6 hours, or 21,600 seconds.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]